Update app.py
I've fixed the FutureWarning in app.py at line 878 by adding the observed=True parameter to the groupby operation. The warning was indicating that the default value of observed=False is deprecated and will change to True in a future version of pandas. By explicitly setting observed=True, we adopt the future default behavior and silence the warning.